prepareMetaForPublish in it's own file, use for publish and pullRedis
sanitize old slugs containing dates
